+++ classes/Overview/class.xaliOverviewGUI.php
@@ -146,6 +146,14 @@ class xaliOverviewGUI extends xaliGUI {
 			$entry->setStatus($status);
 			$entry->setUserId($usr_id);
 			$entry->store();
+            if (intval($status) === xaliChecklistEntry::STATUS_ABSENT_UNEXCUSED) {
+                if (($reason_id = $_POST['absence_reason'][$entry->getId()]) !== null) {
+                    /** @var xaliAbsenceStatement $stm */
+                    $stm = xaliAbsenceStatement::findOrGetInstance($entry->getId());
+                    $stm->setReasonId($reason_id);
+                    $stm->store();
+                }
+            }
 		}
 
 		// update LP
@@ -169,6 +177,14 @@ class xaliOverviewGUI extends xaliGUI {
 			$checklist->setLastEditedBy($this->user->getId());
 			$checklist->setLastUpdate(time());
 			$checklist->store();
+            if (intval($status) === xaliChecklistEntry::STATUS_ABSENT_UNEXCUSED) {
+                if (($reason_id = $_POST['absence_reason'][$entry->getId()]) !== null) {
+                    /** @var xaliAbsenceStatement $stm */
+                    $stm = xaliAbsenceStatement::findOrGetInstance($entry->getId());
+                    $stm->setReasonId($reason_id);
+                    $stm->store();
+                }
+            }
 		}
 
 		// update LP